+++ src/controller/maestrocontroller.h
@@ -7,29 +7,56 @@
 
 #include "core/maestro.h"
 #include "core/section.h"
-#include <memory>
-#include <vector>
+#include "drawingarea/maestrodrawingarea.h"
+#include <QElapsedTimer>
+#include <QObject>
+#include <QSharedPointer>
+#include <QTimer>
+#include <QVector>
 
 using namespace PixelMaestro;
 
-class MaestroController {
+class MaestroDrawingArea;
+
+class MaestroController : public QObject {
+	Q_OBJECT
 	public:
 		MaestroController();
 		~MaestroController();
+		void add_drawing_area(MaestroDrawingArea* drawing_area);
 		Maestro* get_maestro();
+		bool get_running();
 		Show *get_show();
+		uint64_t get_total_elapsed_time();
 		void reset_sections();
 		Section* set_sections(uint8_t num_sections, Point dimensions = Point(10, 10));
+		void stop();
+		void start();
 
 	private:
+		/// References each drawing area that the Maestro is rendering to.
+		QVector<MaestroDrawingArea*> drawing_areas_;
+
+		/// Tracks the time that the Maestro was last paused.
+		uint64_t last_pause_ = 0;
+
+		/// Updates the Maestro's runtime.
+		QElapsedTimer elapsed_timer_;
+
+		/// Timer until the Maestro refreshes.
+		QTimer timer_;
+
 		/// Maestro controlled by this controller.
-		std::shared_ptr<Maestro> maestro_;
+		QSharedPointer<Maestro> maestro_;
 
 		/// The number of Sections in the Maestro.
 		uint8_t num_sections_;
 
 		/// Sections belonging to the Maestro.
 		Section* sections_ = nullptr;
+
+	private slots:
+		void update();
 };
 
 #endif // MAESTROCONTROLLER_H