+++ src/window/settingsdialog.cpp
@@ -37,6 +37,9 @@ SettingsDialog::SettingsDialog(QWidget *parent) : QDialog(parent), ui(new Ui::Se
 	}
 
 	// Check to see if each entry in the output selection box has already been configured.
+	/*
+	 * FIXME: If a device is found in settings but not in the list, add the device to the list, uncheck it, and make it unselectable.
+	 */
 	int num_devices = settings_.beginReadArray(output_devices);
 	for (int list_item = 0; list_item < ui->outputListWidget->count(); list_item++) {
 		for (int device = 0; device < num_devices; device++) {