+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to PixelMaestro Studio will be documented in this file.
 
 The format is loosely based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [v0.11] - In Progress
+### Added
+- Moved Maestro timing to the MaestroController. This should reduce resource usage and prevent multiple outputs from falling out of sync.
+- Animation Editor
+	- Added a Maestro pause/resume button to the Show tab.
+
 ## [v0.10] - 2017-12-05
 ### Added
 - Added donate link to Help menu.